% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellGating.R
\name{addCTLabels}
\alias{addCTLabels}
\title{addCTLabels}
\usage{
addCTLabels(
  cd,
  out,
  outname,
  ct,
  mkrcol = "markers",
  ctcol = "celltype",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cd}{the \code{colData} \code{DataFrame}}

\item{out}{list data structure returned by \code{getCellGroups}}

\item{outname}{character indicating the name of the out data structure}

\item{ct}{character indicating the celltype to assign in the \code{ctcol}}

\item{mkrcol}{character indicating the cd column to store the markers
indicated by \code{outname} (default is markers)}

\item{ctcol}{character indicating the column in cd to store the cell type
indicated by \code{ct} (default is celltype)}

\item{overwrite}{logical indicating if the cell types have to be overwritten
without checking if detected barcodes were already assigned to other celltypes}

\item{verbose}{logical for having informative messages during the execution}
}
\value{
an updated version of the cd DataFrame
}
\description{
addCTLabels
}
