% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITEseq.R
\name{CITEseq}
\alias{CITEseq}
\title{CITEseq}
\usage{
CITEseq(
  DataType = "cord_blood",
  modes = "*",
  version = "1.0.0",
  dry.run = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{DataType}{character(1) indicating the identifier of the dataset to
retrieve.  (default "cord_blood")}

\item{modes}{character( ) The assay types or modes of data to obtain these
include scADT and scRNA-seq data by default.}

\item{version}{character(1) Either version '1.0.0' depending on
data version required.}

\item{dry.run}{logical(1) Whether to return the dataset names before actual
download (default TRUE)}

\item{verbose}{logical(1) Whether to show the dataset currently being
(down)loaded (default TRUE)}

\item{...}{Additional arguments passed on to the
\link[ExperimentHub]{ExperimentHub-class} constructor}
}
\value{
A single cell multi-modal \linkS4class{MultiAssayExperiment} or
    informative `data.frame` when `dry.run` is `TRUE`
}
\description{
function assembles data on-the-fly from `ExperimentHub`
    to provide a \linkS4class{MultiAssayExperiment} container. Actually
    the `dataType` argument provides access to the available datasets
    associated to the package.
}
\details{
CITEseq data are a combination of single cell transcriptomics and
    about a hundread of cell surface proteins.

    Available datasets are:
    \itemize{
        \item{cord_blood: } a dataset of single cells of cord blood as
        provided in Stoeckius et al. (2017).
         \itemize{
            \item{scRNA_Counts} - Stoeckius scRNA-seq gene count matrix
            \item{scADT} - Stoeckius antibody-derived tags (ADT) data
            }
     }
}
\examples{
mse <- CITEseq(dry.run=FALSE)
experiments(mse)
}
\references{
Stoeckius et al. (2017)
}
