% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITEseq.R
\name{.CITEseqMaeToSce}
\alias{.CITEseqMaeToSce}
\title{CITEseqMaeToSce}
\usage{
.CITEseqMaeToSce(mae)
}
\arguments{
\item{mae}{a MultiAssayExperiment object with scRNA and/or scADT and/or 
scHTO named experiments.}
}
\value{
a SingleCellExperiment object as widely with scRNA data as counts
and scADT, scHTO data as altExps.
If only one modality is present, it has returned as main assay of the SCE.
}
\description{
converts a MultiAssayExperiment object with CITEseq data into
a SingleCellExperiment object to be used with already known methods and
packages in literature.
}
\keyword{internal}
