% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontomap.R
\name{ontomap}
\alias{ontomap}
\title{Obtain a map of cell types for each dataset}
\usage{
ontomap(dataset = c("scNMT", "scMultiome", "SCoPE2", "CITEseq", "seqFISH"))
}
\arguments{
\item{dataset}{`character()` One of the existing functions within the
package. If missing, a map of all cell types in each function will
be provided.}
}
\value{
A `data.frame` of metadata regarding the cell types and ontologies,
  if any.
}
\description{
The `ontomap` function provides a mapping of all the cell names across the
all the data sets or for a specified data set.
}
\examples{

ontomap(dataset = "scNMT")

}
