% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMultiModal-package.R
\docType{package}
\name{SingleCellMultiModal-package}
\alias{SingleCellMultiModal-package}
\title{SingleCellMultiModal-package}
\description{
The SingleCellMultiModal package provides a convenient and user-friendly
representation of multi-modal data from project such as \code{scNMT} for mouse
gastrulation.
}
\examples{
help(package = "SingleCellMultiModal")

}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/waldronlab/SingleCellMultiModal/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})
  \item Ricard Argelaguet \email{ricard@ebi.ac.uk}
  \item Dario Righelli \email{dario.righelli@gmail.com}
  \item Kelly Eckenrode \email{kelly.eckenrode@sph.cuny.edu}
  \item Ludwig Geistlinger \email{ludwig_geistlinger@hms.harvard.edu}
  \item Levi Waldron \email{lwaldron.research@gmail.com}
}

Other contributors:
\itemize{
  \item Al Abadi [contributor]
  \item Christophe Vanderaa \email{christophe.vanderaa@uclouvain.be} [contributor]
  \item NCI (GrantNo.: U24CA289073) [funder]
}

}
