% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITEseq.R
\name{CITEseqMseToSce}
\alias{CITEseqMseToSce}
\title{CITEseqMseToSce}
\usage{
CITEseqMseToSce(mse)
}
\arguments{
\item{mse}{a MultiAssayExperiment object with scRNA and scADT named experiments}
}
\value{
a SingleCellExperiment object as widely with scRNA data as counts and
scADT data as altExps
}
\description{
converts a MultiAssayExperiment object with CITEseq data into
a SingleCellExperiment object to be used with already known methods and
packages in literature.
}
\examples{
mse <- CITEseq(dry.run=FALSE)
sce <- CITEseqMseToSce(mse)

}
