% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{scmmCache}
\alias{scmmCache}
\alias{setCache}
\alias{removeCache}
\title{Manage cache / download directories for study data}
\usage{
scmmCache(...)

setCache(
  directory = rappdirs::user_cache_dir("SingleCellMultiModal"),
  verbose = TRUE,
  ask = interactive()
)

removeCache(accession)
}
\arguments{
\item{...}{For \code{scmmCache}, arguments passed to \code{setCache}}

\item{directory}{The file location where the cache is located. Once set
future downloads will go to this folder.}

\item{verbose}{Whether to print descriptive messages}

\item{ask}{logical (default TRUE when interactive session) Confirm the file
location of the cache directory}

\item{cancer_study_id}{A single string from \code{studiesTable} associated
with a study tarball}
}
\description{
Managing data downloads is important to save disk space and
re-downloading data files. This can be done effortlessly via the integrated
\code{BiocFileCache} system.
}
\section{scmmCache}{

Get the directory location of the cache. It will prompt the user to create
a cache if not already created. A specific directory can be used via
\code{setCache}.
}

\section{setCache}{

Specify the directory location of the data cache. By default, it will
got to the user's home/.cache and "appname" directory as specified by
\link{user_cache_dir}. (default appname: 'SingleCellMultiModal')
}

\section{removeCache}{

Some files may become corrupt when downloading, this function allows
the user to delete the tarball associated with a study number in the
cache.
}

