% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMT.R
\name{scNMT}
\alias{scNMT}
\title{Mouse Gastrulation Multi-modal Data}
\source{
\url{http://ftp.ebi.ac.uk/pub/databases/scnmt_gastrulation/}
}
\usage{
scNMT(
  dataType = "mouse_gastrulation",
  modes = "*",
  dry.run = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dataType}{character(1) Indicates study that produces this type of
data (default: 'mouse_gastrulation')}

\item{modes}{character() The assay types or modes of data to obtain these
include single cell Chromatin Accessibilty ("acc"), Methylation ("met"),
RNA-seq ("rna") by default.}

\item{dry.run}{logical(1) Whether to return the dataset names before actual
download (default TRUE)}

\item{verbose}{logical(1) Whether to show the dataset currently being
(down)loaded (default TRUE)}

\item{...}{Additional arguments passed on to the
\link[ExperimentHub]{ExperimentHub} constructor}
}
\value{
A single cell multi-modal \linkS4class{MultiAssayExperiment}
}
\description{
scNMT assembles data on-the-fly from `ExperimentHub` to
    provide a \linkS4class{MultiAssayExperiment} container. The
    `mouse_gastrulation` dataset is obtained from
    \insertCite{Argelaguet2019-et;textual}{SingleCellMultiModal}
}
\examples{
scNMT(dataType = "mouse_gastrulation", modes = "*", dry.run = TRUE)

}
\references{
\insertCite{Argelaguet2019-et}{SingleCellMultiModal}
}
\seealso{
SingleCellMultiModal-package
}
