% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMT.R
\name{scNMT}
\alias{scNMT}
\title{Single-cell Nucleosome, Methylation and Transcription sequencing}
\source{
\url{http://ftp.ebi.ac.uk/pub/databases/scnmt_gastrulation/}
}
\usage{
scNMT(
  DataType = "mouse_gastrulation",
  modes = "*",
  version = "2.0.0",
  dry.run = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{DataType}{character(1) Indicates study that produces this type of
data (default: 'mouse_gastrulation')}

\item{modes}{character() The assay types or modes of data to obtain these
include single cell Chromatin Accessibilty ("acc"), Methylation ("met"),
RNA-seq ("rna") by default.}

\item{version}{character(1) The data version available in ExperimentHub
defaults to the newest version ('2.0.0')}

\item{dry.run}{logical(1) Whether to return the dataset names before actual
download (default TRUE)}

\item{verbose}{logical(1) Whether to show the dataset currently being
(down)loaded (default TRUE)}

\item{...}{Additional arguments passed on to the
\link[ExperimentHub]{ExperimentHub-class} constructor}
}
\value{
A single cell multi-modal \linkS4class{MultiAssayExperiment}
}
\description{
scNMT assembles data on-the-fly from `ExperimentHub` to
    provide a \linkS4class{MultiAssayExperiment} container. The `DataType`
    argument provides access to the `mouse_gastrulation` dataset as obtained
    from Argelaguet et al. (2019; DOI: 10.1038/s41586-019-1825-8).
    Pre-processing code can be seen at
    \url{https://github.com/rargelaguet/scnmt_gastrulation}. Protocol
    information for this dataset is available at Clark et al. (2018). See
    the vignette for the full citation.
}
\details{
scNMT is a combination of RNA-seq (transcriptome) and an adaptation
    of Nucleosome Occupancy and Methylation sequencing (NOMe-seq, the
    methylome and chromatin accessibility) technologies. For more
    information, see Reik et al. (2018) DOI: 10.1038/s41467-018-03149-4

    \itemize{
        \item{mouse_gastrulation:}
        \itemize{
            \item{rna} - RNA-seq
            \item{acc_*} - chromatin accessibility
            \item{met_*} - DNA methylation
            \itemize{
                \item{cgi} - CpG islands
                \item{CTCF} - footprints of CTCF binding
                \item{DHS} - DNase Hypersensitive Sites
                \item{genebody} - gene bodies
                \item{p300} - p300 binding sites
                \item{promoter} - gene promoters
            }
        }
    }
    Special thanks to Al J Abadi for preparing the published data in time
    for the 2020 BIRS Workshop, see the link here:
    url{https://github.com/BIRSBiointegration/Hackathon/tree/master/scNMT-seq}
}
\examples{
scNMT(DataType = "mouse_gastrulation", modes = "*", dry.run = TRUE)

}
\references{
Argelaguet et al. (2019)
}
\seealso{
SingleCellMultiModal-package
}
