% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMultiModal.R
\name{SingleCellMultiModal}
\alias{SingleCellMultiModal}
\title{Combining Modalities into one MultiAssayExperiment}
\usage{
SingleCellMultiModal(
  DataTypes,
  modes = "*",
  versions,
  dry.run = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{DataTypes}{character() A vector of data types as indicated in each
individual function by the \code{DataType} parameter.}

\item{modes}{character() A wildcard / glob pattern of modes, such as
\code{"acc*"}. A wildcard of \code{"*"} will return all modes including
Chromatin Accessibilty ("acc"), Methylation ("met"), RNA-seq ("rna")
which is the default.}

\item{versions}{character() A vector of versions for each DataType.}

\item{dry.run}{logical(1) Whether to return the dataset names before actual
download (default TRUE)}

\item{verbose}{logical(1) Whether to show the dataset currently being
(down)loaded (default TRUE)}

\item{...}{Additional arguments passed on to the
\link[ExperimentHub]{ExperimentHub-class} constructor}
}
\description{
Combine multiple single cell modalities into one using the input of the
individual functions.
}
\examples{

SingleCellMultiModal(c("mouse_gastrulation", "pbmc_10x"), modes = "*",
    dry.run = TRUE, verbose = TRUE
)

}
